perf: add an LRU cache for EVM objects
Adds an LRU cache for different EVM objects for use by
the execution stage.

The cache is intended to be long lived, i.e. it is not cleared
between runs of the execution stage.

The objects cached are:

- Account info
- Storage slots
- Bytecode

Metrics are available for each of the caches:

- Hits
- Misses
- Evictions

This version is very messy; lints are likely to fail and
it is not an ideal cache.

Some improvements to be made:

- Figure out an appropriate sample rate for the W-TinyLFU
  algorithm
- Figure out appropriate cache sizes, including a way to
  approximate memory usage for different cache sizes
  and sampling rates
- Always use the LRU first; if the object is not in the LRU,
  consults the current `PostState`, and if the object
  is not there either, consult the database.

On the last point: we want to use the LRU as much as possible
since this increases precision of the admission filters,
however we cannot *just* rely on the LRU, as state that
has changed (but not yet been committed to disk) may be
evicted at any time. In that case, we need to ensure
that state changes that are not yet committed are still
available in a secondary cache. Since we already store this
data in a cache in `PostState`, it seems appropriate
that this is used.